packages/api/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
firebase-admin==6.3.0
google-cloud-firestore==2.14.0
google-generativeai==0.5.4
Flask-Cors==4.0.0
PyJWT==2.8.0
beautifulsoup4
selenium
webdriver_manager